@@ -4,13 +4,13 @@
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.4.0",
+    "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "axios": "^0.19.2",
-    "bootstrap": "^4.5.3",
-    "react": "^16.12.0",
-    "react-bootstrap": "^1.4.0",
-    "react-dom": "^16.12.0",
+    "bootstrap": "^4.6.0",
+    "react": "^16.14.0",
+    "react-bootstrap": "^1.4.3",
+    "react-dom": "^16.14.0",
     "react-redux": "^7.2.2",
     "react-scripts": "3.4.0",
     "redux": "^4.0.5",
