@@ -1,35 +1,38 @@
-import React, { Component } from "react";
-
+import React, { Component, useReducer, useEffect } from "react";
+import {connect} from 'react-redux'
 import AddForm from './components/AddForm';
 import SmurfDisplay from './components/SmurfDisplay';
-import axios from 'axios';
-
+import {getSmurf} from './actions';
+import {reducer, initialState} from './reducers';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import "./App.css";
 
-class App extends Component {
-  componentDidMount() {
-    axios.get('http://localhost:3333/smurfs').then(res=>{
-      console.log(res);
-    })
-  }
-  
-  render() {
+function App(props) {
+
+const [state, dispatch] = useReducer(reducer, initialState);
+
+useEffect(() => {
+  getSmurf();
+  console.log('test useEffect')
+}, []);
+
+
     return (
       <div className="App">
         <nav className="navbar navbar-dark bg-primary">
-          <a className="navbar-brand">Smurf Village Database</a>
+          <p className="navbar-brand">Smurf Village Database</p>
         </nav>
         <main>
-          <AddForm/>
-          <SmurfDisplay/>
+          <AddForm dispatch={dispatch}/>
+          <SmurfDisplay formData={state} dispatch={dispatch}/>
         </main>
       </div>
-    );
-  }
-}
+    );}
 
-export default App;
+    const mapStateToProps = state => {
+      return {
+        smurf:state.smurfs,
+      }
+    }
 
-//Task List:
-//1. Add in SmurfDisplay and AddForm into your application.
+export default connect(mapStateToProps, {getSmurf})(App);
