@@ -1,14 +1,27 @@
-import React from 'react';
+import React, {useEffect} from 'react';
+import Smurf from './Smurf';
+import {getSmurf} from '../actions';
+import {connect} from 'react-redux';
 
-export class SmurfDisplay extends React.Component {
-    render() {
+const SmurfDisplay = props => {
+    
         return(<div>
-            
+            {props.smurfs.map(item => (
+                <div key={item.id}>
+                    <p>Name: {item.name}</p>
+                    <p>Nickname:{item.nickname}</p>
+                    <p>Position: {item.position}</p>
+                    <p>Description: {item.description}</p>
+                </div>
+            ))}
         </div>)
-    }
+    
+}
+const mapStateToProps = state => {
+    return {smurfs:state.smurfs}
 }
 
-export default SmurfDisplay;
+export default connect(mapStateToProps, {getSmurf})(SmurfDisplay);
 
 //Task List:
 //1. Import in all needed components and library methods.
